// backend/src/domain/financial/cash-flow-statement.types.ts
export interface CashFlowStatement {
  date: string;

  operatingCashFlow: number;

 capitalExpenditure: number;

 freeCashFlow: number;

 investmentsInPropertyPlantAndEquipment: number;

 netCashProvidedByOperatingActivities: number;

 netCashUsedForInvestingActivities: number;

 netCashUsedProvidedByFinancingActivities: number;

 netChangeInCash: number;
}